requirements.txt
beautifulsoup4==4.6.0
Django==2.1.4
django-filebrowser==3.11.1
django-grappelli==2.12.1
facebook-sdk
# google-api-python-client
html2text
httplib2
# Upgrading this will require a painful time-consuming
# change to get most popular articles from Google Analytics.
oauth2client
oauthlib
Pillow
psycopg2-binary
pycrypto
pytz
requests
requests-oauthlib
sqlparse
tweepy
Unidecode
uritemplate
wheel
django-haystack
python-dateutil
piexif
django-allauth
smartypants
django-ajax-selects
django-debug-toolbar
django-el-pagination
csscompressor
django-compressor
django-allauth-2fa
